Final validation Identified three errors in the code, and one alternative approach that led to consistency with validation test cases.
Bugs fixed:
The definition of NB and NBi_temp were incorrect, and led to incorrect averaging of the lowest 3 critical bands.
There was an error in the calculation of Nz_basis, for which the Boolean mask was incorrectly defined.
Indexing of the Phik_lztau (the DFT of the lag-windowed ACF) was incorrectly limited to the block size s_b(z), which led to an incorrect calculation of the max for the specific tonal loudness.
The use of resample to align time-dependent variable sampling sizes was causing rippling in the calculated values, which led to incorrect estimation of specific noise loudness. The use of resample has been replaced by interp1.

Features added:
A new input variable 'ecma' has been defined to allow the user to decide whether to maintain strict adherence to Equation 40 of the standard, or whether to compensate the output for the signal start zero-padding of the signal at the start - the standard currently only compensates for the zero-padding at the signal end.